Added Refactoring Added checks for expected output dataframe format
