+++ hca2scea-backend/json_files/sdrf_map.json
@@ -1,0 +1,13 @@
+{"Characteristics[organism]": "donor_organism.genus_species.ontology_label",
+"Characteristics[individual]": "donor_organism.biomaterial_core.biomaterial_id",
+"Characteristics[sex]": "donor_organism.sex",
+"Characteristics[age]": "donor_organism.organism_age",
+"Unit[time unit]": "donor_organism.organism_age_unit.text",
+"Characteristics[developmental stage]": "donor_organism.development_stage.text",
+"Characteristics[organism part]": "specimen_from_organism.organ.ontology_label",
+"Characteristics[sampling site]": "specimen_from_organism.organ_parts.ontology_label",
+"Characteristics[cell type]": "cell_suspension.selected_cell_types.ontology_label",
+"Characteristics[disease]": "donor_organism.diseases.ontology_label",
+"Characteristics[organism status]": "donor_organism.is_living",
+"Characteristics[cause of death]": "donor_organism.death.cause_of_death",
+"Description": "specimen_from_organism.biomaterial_core.biomaterial_description"}
