--- hca2scea-backend/test/golden/output/Test01/E-HCAD-E-HCAD-50.sdrf.txt
@@ -1,4 +1,0 @@
-Source Name	Characteristics[organism]	Characteristics[individual]	Characteristics[sex]	Characteristics[age]	Unit [time unit]	Characteristics[developmental stage]	Characteristics[organism part]	Characteristics[sampling site]	Characteristics[cell type]	Characteristics[disease]	Characteristics[organism status]	Characteristics[cause of death]	Characteristics[clinical history]	Description	Material Type	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Extract Name	Material Type	Comment[library construction]	Comment[input molecule]	Comment[primer]	Comment[end bias]	Comment[umi barcode read]	Comment[umi barcode offset]	Comment[umi barcode size]	Comment[cell barcode read]	Comment[cell barcode offset]	Comment[cell barcode size]	Comment[sample barcode read]	Comment[sample barcode offset]	Comment[sample barcode size]	Comment[single cell isolation]	Comment[cDNA read]	Comment[cDNA read offset]	Comment[cDNA read size]	Comment[LIBRARY_STRAND]	Comment[LIBRARY_LAYOUT]	Comment[LIBRARY_SOURCE]	Comment[LIBRARY_STRATEGY]	Comment[LIBRARY_SELECTION]	Protocol REF	Assay Name	Technology Type	Scan Name	Comment[RUN]	Comment[BioSD_SAMPLE]	Comment[ENA_EXPERIMENT]	Comment [BioSD_SAMPLE]	Comment[ENA_RUN]
-GSM4486714	Homo sapiens	AXH009	unknown	24 to 57	year	human adult stage	testis	epididymis		testicular cancer	alive			human epididymis epithelial tissue [AXH009]	cell	P-HCADE-HCAD-50-1	P-HCADE-HCAD-50-2	P-HCADE-HCAD-50-3	P-HCADE-HCAD-50-4	P-HCADE-HCAD-50-5	P-HCADE-HCAD-50-6	P-HCADE-HCAD-50-7	P-HCADE-HCAD-50-8	GSM4486714	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-50-9	GSM4486714	sequencing assay	GSM4486714	GSM4486714	SAMN14646191		SAMN14646191	SRR11575369
-GSM4486715	Homo sapiens	AXH0012	unknown	24 to 57	year	human adult stage	testis	epididymis		testicular cancer	alive			human epididymis epithelial tissue [AXH012]	cell	P-HCADE-HCAD-50-1	P-HCADE-HCAD-50-2	P-HCADE-HCAD-50-3	P-HCADE-HCAD-50-4	P-HCADE-HCAD-50-5	P-HCADE-HCAD-50-6	P-HCADE-HCAD-50-7	P-HCADE-HCAD-50-8	GSM4486715	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-50-9	GSM4486715	sequencing assay	GSM4486715	GSM4486715	SAMN14646190		SAMN14646190	SRR11575370
-GSM4486716	Homo sapiens	AXH0014	unknown	24 to 57	year	human adult stage	testis	epididymis		testicular cancer	alive			human epididymis epithelial tissue [AXH014]	cell	P-HCADE-HCAD-50-1	P-HCADE-HCAD-50-2	P-HCADE-HCAD-50-3	P-HCADE-HCAD-50-4	P-HCADE-HCAD-50-5	P-HCADE-HCAD-50-6	P-HCADE-HCAD-50-7	P-HCADE-HCAD-50-8	GSM4486716	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-50-9	GSM4486716	sequencing assay	GSM4486716	GSM4486716	SAMN14646189		SAMN14646189	SRR11575371
