--- hca2scea-backend/test/golden/output/Test02/E-HCAD-E-HCAD-52.idf.txt
@@ -1,30 +1,0 @@
-
-MAGE-TAB Version	1.1
-Investigation Title	Transcriptomic Profiling of the Developing Cardiac Conduction System at Single-Cell Resolution.
-Comment[Submitted Name]	DevelopingCardiacSystem
-Experiment Description	The cardiac conduction system (CCS) consists of distinct components including the sinoatrial node (SAN), atrioventricular node (AVN), His bundle, bundle branches (BB) and Purkinje fibers (PF). Despite an essential role for the CCS in heart development and function, the CCS has remained challenging to interrogate due to inherent obstacles including small cell numbers, large cell type heterogeneity, complex anatomy and difficulty in isolation. We used single-cell RNA-sequencing (scRNA-seq) to perform genome-wide analysis of gene expression of the CCS at single-cell resolution. Overall design: Single cell transcriptional analysis of embryonic cardiac cells using 10x genomics 3' solution.
-Public Release Date	2019-08-01
-Person First Name	William	Benjamin	David	Lei	Guang	Jan	Orlando	 Shannon	Sneha	Eliza	Marc	Joseph	Sean	Ami
-Person Last Name	Goodyer	Beyersdorf	Paik	Tian	Li	Buikema	Chirikian	Choi	Venkatraman	Adams	Tessier-Lavigne	Wu	Wu 	Day
-Person Mid Initials	R	M	T			W				L		C	M	E
-Person Email	wgoodyer@stanford.edu												smwu@stanford.edu	ami@ebi.ac.uk
-Person Affiliation	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University School of Medicine	Stanford University	Stanford University	Stanford University School of Medicine	Stanford University School of Medicine	EMBL-EBI
-Person Address	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	386A Noe Street,CA,94114	371 Serra Mall, Stanford, CA 94305	371 Serra Mall, Stanford, CA 94305	386A Noe Street,CA,94114	386A Noe Street,CA,94114	EMBL-EBI,Wellcome Genome campus, Hinxton
-Person Roles	submitter												submitter	data curator
-Protocol Type	sample collection protocol	enrichment protocol	nucleic acid library construction protocol	nucleic acid sequencing protocol
-Protocol Name	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	P-HCADE-HCAD-52-4
-Protocol Description	Wild-type, timed pregnant CD1 mice were acquired from Jackson Laboratory (Sacramento, CA, USA). Embryonic day 16.5 (E16.5), wild-type CD1 mice were obtained from pregnant mice. Embryonic mouse hearts were harvested, and 3 zones of microdissection were isolated based on anatomic landmarks and entailed: Zone I—SAN region (superior vena cava/right atrial junction), Zone II—AVN/His region (crux of heart), and Zone III—BB/PF region (luminal side of ventricles). Specifically, Zone II was dissected as a large area at the crux of the heart from the base of the interatrial sep- tum (including the triangle of Koch) to below the plane of the mitral annulus, from the posterior-most aspect of the heart to the anterior-most. Tissues from a total of 6 different embryos were pooled for each zone of dissection. 	Dissected tissues were dissociated into single cells in a microcentrifuge tube with 100ul of 0.25% trypsin and incubated at 37 degrees celcius for 10 min. Subsequently, 1.4 mL of collagenase A/B (10 mg/ml) and 20% FBS serum in HBSS was added to the micfocentrifuge tubes and tissue samples were returned to 37 degrees celcius in a waterbath for an additional 20 min. Cells were then spun down at 1000 RPM for 5 min, supernatant was removed and cells werw washed in 20% FBS in HBSS. After suspension, cells were resuspended to around 600 cell/ul with 0.04% FBS/HBSS solution for processing on the 10X platform.	Prepared cells were captured with 10X Chromium by following the Chromium single-cell 3' reagent kits v2 user guide. Briefly, single cells were partitioned into nanoliter-scale Gel Bread-In-Emulsions in the Chromium controller. After dissolution of the Gel beads in GEMS, the primers were released, and mRNA were reverse transcribed into a barcoded cDNA library. After further clean-up and amplification, the cDNA was enzymatically fragmented and 3' end fragments were selected for library preparation. After further end repair, A-tailing, adaptor ligation and PCR amplification, sample index, UMI sequences, barcode sequences and sequencing primer P5 and P7 on both ends were added to cDNA.	Libraries were sequenced using the Illumina HiSeq 4000 instrument.
-Protocol Hardware				Illumina HiSeq 4000
-Term Source Name	EFO	ArrayExpress
-Term Source File	http://www.ebi.ac.uk/efo/efo.owl	http://www.ebi.ac.uk/arrayexpress/
-Comment[AEExperimentType]	RNA-seq of coding RNA from single cells
-Experimental Factor Name	organ	part
-Experimental Factor Type	organ	part
-Comment[EAAdditionalAttributes]
-Comment[EACurator]	AD
-Comment[EAExpectedClusters]	
-Comment[ExpressionAtlasAccession]	E-HCAD-E-HCAD-52
-Comment[HCALastUpdateDate]	2021-03-01
-Comment[SecondaryAccession]	2086eb05-10b9-432b-b7f0-169ccc49d270	GSE132658
-Comment[EAExperimentType]	differential
-SDRF File	E-HCAD-E-HCAD-52.sdrf.txt
