--- hca2scea-backend/test/golden/output/Test02/E-HCAD-E-HCAD-52.sdrf.txt
@@ -1,9 +1,0 @@
-Source Name	Characteristics[organism]	Characteristics[individual]	Characteristics[sex]	Characteristics[age]	Unit [time unit]	Characteristics[developmental stage]	Characteristics[organism part]	Characteristics[sampling site]	Characteristics[cell type]	Characteristics[disease]	Characteristics[organism status]	Characteristics[cause of death]	Characteristics[clinical history]	Description	Material Type	Protocol REF	Protocol REF	Protocol REF	Extract Name	Material Type	Comment[library construction]	Comment[input molecule]	Comment[primer]	Comment[end bias]	Comment[umi barcode read]	Comment[umi barcode offset]	Comment[umi barcode size]	Comment[cell barcode read]	Comment[cell barcode offset]	Comment[cell barcode size]	Comment[sample barcode read]	Comment[sample barcode offset]	Comment[sample barcode size]	Comment[single cell isolation]	Comment[cDNA read]	Comment[cDNA read offset]	Comment[cDNA read size]	Comment[LIBRARY_STRAND]	Comment[LIBRARY_LAYOUT]	Comment[LIBRARY_SOURCE]	Comment[LIBRARY_STRATEGY]	Comment[LIBRARY_SELECTION]	Protocol REF	Assay Name	Technology Type	Scan Name	Comment[RUN]	Comment[BioSD_SAMPLE]	Comment[ENA_EXPERIMENT]	Comment [BioSD_SAMPLE]	Comment[ENA_RUN]
-SRX6060231	Mus musculus	mouse_donor_1	unknown	16.5	day	Theiler Stage 24	heart	atrioventricular node||atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone II AVN/His region	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060231	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060231	sequencing assay	SRX6060231	SRX6060231	SAMN12046032		SAMN12046032	SRR9290713
-SRX6060231	Mus musculus	mouse_donor_1	unknown	16.5	day	Theiler Stage 24	heart	atrioventricular node||atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone II AVN/His region	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060231	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060231	sequencing assay	SRX6060231	SRX6060231	SAMN12046032		SAMN12046032	SRR9290714
-SRX6060230	Mus musculus	mouse_donor_1	unknown	16.5	day	Theiler Stage 24	heart	sinoatrial node		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone I SAN region	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060230	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060230	sequencing assay	SRX6060230	SRX6060230	SAMN12046033		SAMN12046033	SRR9290711
-SRX6060230	Mus musculus	mouse_donor_1	unknown	16.5	day	Theiler Stage 24	heart	sinoatrial node		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone I SAN region	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060230	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060230	sequencing assay	SRX6060230	SRX6060230	SAMN12046033		SAMN12046033	SRR9290712
-SRX6060233	Mus musculus	mouse_donor_2	unknown	16.5	day	Theiler Stage 24	heart	heart right ventricle||right branch of atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone III BB/PF region (Right Ventricle)	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060233	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060233	sequencing assay	SRX6060233	SRX6060233	SAMN12046030		SAMN12046030	SRR9290717
-SRX6060233	Mus musculus	mouse_donor_2	unknown	16.5	day	Theiler Stage 24	heart	heart right ventricle||right branch of atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone III BB/PF region (Right Ventricle)	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060233	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060233	sequencing assay	SRX6060233	SRX6060233	SAMN12046030		SAMN12046030	SRR9290718
-SRX6060232	Mus musculus	mouse_donor_2	unknown	16.5	day	Theiler Stage 24	heart	heart left ventricle||left branch of atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone III BB/PF region (Left Ventricle)	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060232	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060232	sequencing assay	SRX6060232	SRX6060232	SAMN12046031		SAMN12046031	SRR9290715
-SRX6060232	Mus musculus	mouse_donor_2	unknown	16.5	day	Theiler Stage 24	heart	heart left ventricle||left branch of atrioventricular bundle		normal	dead			Microdissected region from embryonic day 16.5 (E16.5) mouse hearts; Zone III BB/PF region (Left Ventricle)	cell	P-HCADE-HCAD-52-1	P-HCADE-HCAD-52-2	P-HCADE-HCAD-52-3	SRX6060232	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-52-4	SRX6060232	sequencing assay	SRX6060232	SRX6060232	SAMN12046031		SAMN12046031	SRR9290716
