--- hca2scea-backend/test/golden/output/Test03/E-HCAD-E-HCAD-54.idf.txt
@@ -1,30 +1,0 @@
-
-MAGE-TAB Version	1.1
-Investigation Title	Transcriptomic and clonal characterization of T cells in the human central nervous system.
-Comment[Submitted Name]	TCellsHumanCentralNervousSystem
-Experiment Description	Tissues develop unique homeostatic immune states through specific recruitment of immune cells that are further shaped by the tissue environment. Single-cell RNA of healthy cerebrospinal fluid (CSF) and single-nucleus RNA sequencing of brain parenchyma was used to profile the T cell state in the human central nervous system (CNS).
-Public Release Date	2020-09-02
-Person First Name	Jenna	Le	Maggie	Kelly	Chrysoula	Khadir	Ahmad	Smita	Jack	David	David	Ami
-Person Last Name	Pappalardo	Zhang	Pecsok	Perlman	Zografou	Raddassi	Abulaban	Krishnaswamy	Antel	van Dijk	Hafler	Day
-Person Mid Initials	L		K								A	
-Person Email										david.vandijk@yale.edu	david.hafler@yale.edu	
-Person Affiliation	Yale School of Medicine	Yale School of Medicine	Yale School of Medicine	Montreal Neurologic Institute	Yale School of Medicine	Yale School of Medicine	Yale School of Medicine	Yale School of Medicine	Montreal Neurologic Institute	Yale School of Medicine	Yale School of Medicine	EMBL-EBI
-Person Address	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Montreal, Quebec	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Montreal, Quebec	Yale School of Medicine, New Haven, CT 06511	Yale School of Medicine, New Haven, CT 06511	Wellcome Genome Campus, Hinxton
-Person Roles										submitter	submitter	
-Protocol Type	sample collection protocol	sample collection protocol	enrichment protocol	enrichment protocol	enrichment protocol	enrichment protocol	nucleic acid library construction protocol	nucleic acid sequencing protocol
-Protocol Name	P-HCADE-HCAD-54-1	P-HCADE-HCAD-54-2	P-HCADE-HCAD-54-3	P-HCADE-HCAD-54-4	P-HCADE-HCAD-54-5	P-HCADE-HCAD-54-6	P-HCADE-HCAD-54-7	P-HCADE-HCAD-54-8
-Protocol Description	Flash-frozen human brain tissue was obtained from epilepsy surgery of temporal lobectomy in the Department of Neurosurgery, Yale School of Medicine. Nuclei were isolated from flash-frozen human brain tissue as previously described (77) with modifications.	Post-mortem unfixed frozen human brain tissues were obtained from the Department of Pathology, Yale School of Medicine. Nuclei were isolated from post-mortem frozen human brain tissue as previously described (77) with modifications.					single-nucleus RNA-seq libraries were prepared using the Chromium Single Cell 3′ Reagent Kit v3 chemistry according to the manufacturer’s instructions (10x Genomics).	Single-cell RNA sequencing libraries were sequenced on an Illumina HiSeq 4000 instrument.
-Protocol Hardware								Illumina HiSeq 4000
-Term Source Name	EFO	ArrayExpress
-Term Source File	http://www.ebi.ac.uk/efo/efo.owl	http://www.ebi.ac.uk/arrayexpress/
-Comment[AEExperimentType]	RNA-seq of coding RNA from single cells
-Experimental Factor Name	organ	part
-Experimental Factor Type	organ	part
-Comment[EAAdditionalAttributes]
-Comment[EACurator]	AD
-Comment[EAExpectedClusters]	
-Comment[ExpressionAtlasAccession]	E-HCAD-E-HCAD-54
-Comment[HCALastUpdateDate]	2021-03-01
-Comment[SecondaryAccession]	d3a4ceac-4d66-4984-9704-2570c0647a56	GSE157277	SRP279608 
-Comment[EAExperimentType]	differential
-SDRF File	E-HCAD-E-HCAD-54.sdrf.txt
