--- hca2scea-backend/test/golden/output/Test04/E-HCAD-E-HCAD-55.sdrf.txt
@@ -1,10 +1,0 @@
-Source Name	Characteristics[organism]	Characteristics[individual]	Characteristics[sex]	Characteristics[age]	Unit [time unit]	Characteristics[developmental stage]	Characteristics[organism part]	Characteristics[sampling site]	Characteristics[cell type]	Characteristics[disease]	Characteristics[organism status]	Characteristics[cause of death]	Characteristics[clinical history]	Description	Material Type	Protocol REF	Protocol REF	Protocol REF	Protocol REF	Extract Name	Material Type	Comment[library construction]	Comment[input molecule]	Comment[primer]	Comment[end bias]	Comment[umi barcode read]	Comment[umi barcode offset]	Comment[umi barcode size]	Comment[cell barcode read]	Comment[cell barcode offset]	Comment[cell barcode size]	Comment[sample barcode read]	Comment[sample barcode offset]	Comment[sample barcode size]	Comment[single cell isolation]	Comment[cDNA read]	Comment[cDNA read offset]	Comment[cDNA read size]	Comment[LIBRARY_STRAND]	Comment[LIBRARY_LAYOUT]	Comment[LIBRARY_SOURCE]	Comment[LIBRARY_STRATEGY]	Comment[LIBRARY_SELECTION]	Protocol REF	Assay Name	Technology Type	Scan Name	Comment[RUN]	Comment[BioSD_SAMPLE]	Comment[ENA_EXPERIMENT]	Comment [BioSD_SAMPLE]	Comment[ENA_RUN]
-SRX5255152	Homo sapiens	donor_1	male	41.0	year	human adult stage	blood		leukocyte		dead	brain palsy		donor 1 blood	cell	P-HCADE-HCAD-55-1	P-HCADE-HCAD-55-4	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255152	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255152	sequencing assay	SRX5255152	SRX5255152	SAMN10755966		SAMN10755966	SRR8448139
-SRX5255153	Homo sapiens	donor_1	male	41.0	year	human adult stage	spleen		leukocyte		dead	brain palsy		donor 1 spleen	cell	P-HCADE-HCAD-55-2	P-HCADE-HCAD-55-5	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255153	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255153	sequencing assay	SRX5255153	SRX5255153	SAMN10755965		SAMN10755965	SRR8448140
-SRX5255154	Homo sapiens	donor_1	male	41.0	year	human adult stage	liver		leukocyte		dead	brain palsy		donor 1 liver	cell	P-HCADE-HCAD-55-3	P-HCADE-HCAD-55-6	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255154	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255154	sequencing assay	SRX5255154	SRX5255154	SAMN10755964		SAMN10755964	SRR8448141
-SRX5255155	Homo sapiens	donor_2	male	27.0	year	human adult stage	blood		leukocyte		dead	brain palsy		donor 2 blood	cell	P-HCADE-HCAD-55-1	P-HCADE-HCAD-55-4	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255155	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255155	sequencing assay	SRX5255155	SRX5255155	SAMN10755963		SAMN10755963	SRR8448142
-SRX5255156	Homo sapiens	donor_2	male	27.0	year	human adult stage	spleen		leukocyte		dead	brain palsy		donor 2 spleen	cell	P-HCADE-HCAD-55-2	P-HCADE-HCAD-55-5	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255156	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255156	sequencing assay	SRX5255156	SRX5255156	SAMN10755962		SAMN10755962	SRR8448143
-SRX5255157	Homo sapiens	donor_2	male	27.0	year	human adult stage	liver		leukocyte		dead	brain palsy		donor 2 liver	cell	P-HCADE-HCAD-55-3	P-HCADE-HCAD-55-6	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255157	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255157	sequencing assay	SRX5255157	SRX5255157	SAMN10755961		SAMN10755961	SRR8448144
-SRX5255158	Homo sapiens	donor_3	male	56.0	yeay	human adult stage	blood		leukocyte		dead	brain palsy		donor 3 blood	cell	P-HCADE-HCAD-55-1	P-HCADE-HCAD-55-4	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255158	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255158	sequencing assay	SRX5255158	SRX5255158	SAMN10755960		SAMN10755960	SRR8448145
-SRX5255159	Homo sapiens	donor_3	male	56.0	yeay	human adult stage			leukocyte		dead	brain palsy		donor 3 spleen	cell	P-HCADE-HCAD-55-2	P-HCADE-HCAD-55-5	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255159	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255159	sequencing assay	SRX5255159	SRX5255159	SAMN10755959		SAMN10755959	SRR8448146
-SRX5255160	Homo sapiens	donor_3	male	56.0	yeay	human adult stage	liver		leukocyte		dead	brain palsy		donor 3 liver	cell	P-HCADE-HCAD-55-3	P-HCADE-HCAD-55-6	P-HCADE-HCAD-55-7	P-HCADE-HCAD-55-8	SRX5255160	RNA	10xV2	polyA RNA	oligo-dT	3 prime tag	Read 1	16	10	read1	0	10		0	8	10X	read2	0	98	first strand	PAIRED	TRANSCRIPTOMIC SINGLE CELL	RNA-Seq	cDNA	P-HCADE-HCAD-55-9	SRX5255160	sequencing assay	SRX5255160	SRX5255160	SAMN10755958		SAMN10755958	SRR8448147
